fix(core,loop): move the merge-signal reader beside its own signal (#4044)
CI caught a real edge, not a lint nit. `read_merge_signal` lived in
`teatree.loop.admission`, so `scanner_factories` reached it with a
function-scoped `from teatree.loop... import` — an intra-loop deferred edge,
which is exactly the shape tach's acyclic guard cannot see. The ratchet
(`test_intra_loop_deferred_import_ratchet`) blocked it and offered two ways out:
declare the edge, or bump the peg.

Neither was right, because the placement was. `read_quota_signal` already lives
in `core/admission_governor.py` and does its ORM read there; the merge reader is
the same kind of thing and belongs beside the signal it reads, not one package
away. Moving it removes the intra-loop edge entirely rather than declaring or
excusing it, and `scanner_factories` now imports from core like its neighbours.

That does add a second call-time ORM import to `admission_governor.py`, which is
the peg this bumps 1 -> 2 — the app registry is not ready at module import, so
`PullRequest` / `SweepSkipStreak` can no more be top-level than
`AnthropicTokenUsage` could. Rationale recorded on the peg entry itself.
